Move cursednesslevel enum to client ui folder
