+++ core/src/mindustry/ui/fragments/MenuFragment.java
@@ -14,7 +14,7 @@ import arc.scene.ui.TextButton.*;
 import arc.scene.ui.layout.*;
 import arc.struct.*;
 import arc.util.*;
-import mindustry.client.CursednessLevel;
+import mindustry.client.ui.CursednessLevel;
 import mindustry.client.ui.*;
 import mindustry.core.*;
 import mindustry.game.EventType.*;
